requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.4
django-dynamic-fixtures==0.2.1
django-rest-swagger==2.2.0
djangorestframework==3.15.1
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==25.0.1
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
psycopg==3.1.19
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pyflakes==3.2.0
pytest==8.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
